feat: support for redwood (#9)
* allow MFEs to be hosted on either the LMS or the CMS

Before all the MFEs that were hosted by path were hosted in the LMS
domain. For certain MFEs (course-authoring) the CMS domain is more
appropriate.

* point the MFEs URLs to either the LMS or the CMS

* handle special cases in the $LMS_HOST/account route.

The `/account/password` endpoint is used in the 'forgot password' flow,
for that reason we must forward those requests to the LMS.

The `/account/settings` is more specific to the multitenant use
case. Some sites may be configured to use the legacy pages but the
current configuration would send every subpath of `/account/` to the
MFE, we make an exception for the legacy URL.