fix: load the configuration using an Action hook (#17)
When configuring MFEs via config.yml file the tutor configuration was
loaded by setting the tutor root to the current directory. This caused
an error whenever `tutor config save` was ran outside the root
directory.

The configuration is now loaded via the CONFIG_LOADED Action hook before
registering the callback that registers the final MFEs in the MFE_APPS
hook.